requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.14.1
six==1.16.0
